\name{EricsOtherLinePickingAllProblems}
\alias{EricsOtherLinePickingAllProblems}
\title{Gets list of all problems}
\usage{
  EricsOtherLinePickingAllProblems()
}
\arguments{
  \item{short_tab}{boolean is false gives description as
  well.}
}
\value{
  Data frame with information on each problem
}
\description{
  Problems are integers describing possible geometries for
  the line picking problem. This function prints all
  implemented problems.
}
\note{
  September 8 2012
}
\examples{
EricsOtherLinePickingAllProblems()
}
\author{
  Eric Parsonage
}

