\name{FindProblem}
\alias{FindProblem}
\title{Gets problem from list of problems}
\usage{
  FindProblem(x, problem = 0)
}
\arguments{
  \item{x}{list of all problems}

  \item{problem}{problem number or name}
}
\value{
  index or -1 if error
}
\description{
  Takes either a problem number of name and finds it in the
  list of all problems
}
\note{
  September 14 2012
}
\author{
  Eric Parsonage, Matt Roughan, Jono Tuke
}
\keyword{internal}

