\name{FormatProblem}
\alias{FormatProblem}
\title{Clean printing of problem}
\usage{
  FormatProblem(x)
}
\arguments{
  \item{x}{problem list}
}
\value{
  prints out information nice-like
}
\description{
  Nice printing of problem list
}
\author{
  Eric Parsonage, Matt Roughan, Jono Tuke
}
\keyword{internal}

