\name{GetNumCDF}
\alias{GetNumCDF}
\title{Get the numeric cdf for LinePicking problem}
\usage{
  GetNumCDF(problem, parameters, t)
}
\arguments{
  \item{problem}{the problem number}

  \item{parameters}{the parameters of the problem of
  interest}

  \item{t}{value for cdf}
}
\value{
  cdf
}
\description{
  Uses numerical integration to give the cdf
}
\note{
  September 14 2012
}
\author{
  Eric Parsonage, Matt Roughan, Jono Tuke
}
\keyword{internal}

