\name{GetNumMoment}
\alias{GetNumMoment}
\title{Get the numeric kth moment for LinePicking problem}
\usage{
  GetNumMoment(problem, parameters, k = 1)
}
\arguments{
  \item{problem}{the problem number}

  \item{parameters}{the parameters of the problem of
  interest}

  \item{k}{which moment}
}
\value{
  expected value
}
\description{
  Uses numerical integration to give the kth moment
}
\note{
  September 14 2012
}
\author{
  Eric Parsonage, Matt Roughan, Jono Tuke
}
\keyword{internal}

