\name{linepdf-package}
\alias{linepdf-package}
\alias{linepdf}
\docType{package}
\title{
Give probability density function for distance between to Poisson points in space
}
\description{
Takes a vector of points and gives probability density function for those distances
}
\details{
\tabular{ll}{
Package: \tab linepdf\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2012-08-17\cr
License: \tab GPL 3.0\cr
}
}
\author{
Matt Roughan, Jono Tuke, Eric Parsonage

Maintainer: Who to complain to <simon.tuke@adelaide.edu.au>
}
\references{
Matt and Eric paper
}
\keyword{ package }
\examples{
GetDistance(t=1:100,mode=0,para=50)
}
