\name{LinePickingAllModes}
\alias{LinePickingAllModes}
\title{Gets list of all modes}
\usage{
  LinePickingAllModes()
}
\value{
  List with \itemize{ \item names: list of the spaces names
  \item description: brief description of each space }
}
\description{
  Modes are integers describing possible geometries for the
  line picking problem. This function prints all possible
  modes.
}
\note{
  August 25 2012
}
\examples{
LinePickingAllModes()
}
\author{
  Eric Parsonage, Matt Roughan, Jono Tuke
}

