\name{LinePickingAllProblems}
\alias{LinePickingAllProblems}
\title{Gets list of all problems}
\usage{
  LinePickingAllProblems()
}
\value{
  List with \itemize{ \item names: list of the spaces names
  \item description: brief description of each space }
}
\description{
  Problemss are integers describing possible geometries for
  the line picking problem. This function prints all
  possible problems.
}
\note{
  August 25 2012
}
\examples{
LinePickingAllProblems()
}
\author{
  Eric Parsonage, Matt Roughan, Jono Tuke
}

