\name{LinePickingModeLookup}
\alias{LinePickingModeLookup}
\title{Gets information for a given mode}
\usage{
  LinePickingModeLookup(mode = 0)
}
\arguments{
  \item{mode}{integer from 0 to 5}
}
\value{
  prints out information
}
\description{
  Modes are integers describing possible geometries for the
  line picking problem. This function gives you the
  information for a given mode.
}
\note{
  August 25 2012
}
\examples{
LinePickingModeLookup(mode=0)
}
\author{
  Matt Roughan, Jono Tuke, Eric Parsonage
}

