\name{LinePickingNparLookup}
\alias{LinePickingNparLookup}
\title{Gets number of problems}
\usage{
  LinePickingNparLookup(problem = 1)
}
\arguments{
  \item{problem}{number of problem of interest}
}
\value{
  N number of parameters for this problem
}
\description{
  Problems are integers describing possible geometries for
  the line picking problem.
}
\examples{
LinePickingProblemNparLookup(problem=1)
}
\author{
  Eric Parsonage, Matt Roughan, Jono Tuke
}

