\name{LinePickingNumberOfProblems}
\alias{LinePickingNumberOfProblems}
\title{Gets number of problems}
\usage{
  LinePickingNumberOfProblems()
}
\value{
  number of problems
}
\description{
  Problems are integers describing possible geometries for
  the line picking problem.
}
\examples{
LinePickingNumberOfProblems()
}
\author{
  Eric Parsonage, Matt Roughan, Jono Tuke
}

