\name{LinePickingProblemLookup}
\alias{LinePickingProblemLookup}
\title{Function to print out information about a given problem}
\usage{
  LinePickingProblemLookup(problem = 0, print = TRUE)
}
\arguments{
  \item{problem}{see \code{\link{LinePickingPDF}}}

  \item{print}{boolean if TRUE prints out problem nicely}
}
\value{
  list of problem info
}
\description{
  Uses internal function format list to print nice
}
\examples{
LinePickingProblemLookup(problem=1)
}
\author{
  Eric Parsonage, Matt Roughan, Jono Tuke
}

