\name{LinePickingAllmodes}
\alias{LinePickingAllmodes}
\title{Prints list of all modes}
\usage{
  LinePickingAllmodes()
}
\value{
  prints out information
}
\description{
  Modes are integers describing possible geometries for the
  line picking problem. This function prints all possible
  modes.
}
\note{
  August 25 2012
}
\examples{
LinePickingAllmodes()
}
\author{
  Matt Roughan, Jono Tuke, Eric Parsonage
}

