\name{LinePickingAllModes}
\alias{LinePickingAllModes}
\title{Prints list of all modes}
\usage{
  LinePickingAllModes()
}
\value{
  prints out information
}
\description{
  Modes are integers describing possible geometries for the
  line picking problem. This function prints all possible
  modes.
}
\note{
  August 25 2012
}
\examples{
LinePickingAllModes()
}
\author{
  Eric Parsonage, Matt Roughan, Jono Tuke
}

