\name{LinePickingAllProblems}
\alias{LinePickingAllProblems}
\title{Gets list of all problems}
\usage{
  LinePickingAllProblems(print = TRUE)
}
\arguments{
  \item{print}{boolean if TRUE then prints out all the
  info}
}
\value{
  list with information on each problem
}
\description{
  Problems are integers describing possible geometries for
  the line picking problem. This function prints all
  possible problems.
}
\examples{
LinePickingAllProblems()
}
\author{
  Eric Parsonage, Matt Roughan, Jono Tuke
}

