\name{LinePickingProblemLookup}
\alias{LinePickingProblemLookup}
\title{Gets information for a given problem}
\usage{
  LinePickingProblemLookup(problem = 0)
}
\arguments{
  \item{problem}{see \code{\link{LinePickingPDF}}}
}
\value{
  prints out information
}
\description{
  Problemss are integers describing possible geometries for
  the line picking problem. This function gives you the
  information for a given problem.
}
\note{
  August 25 2012
}
\examples{
LinePickingProblemLookup(problem=0)
}
\author{
  Eric Parsonage, Matt Roughan, Jono Tuke
}

