\name{LinePickingVar}
\alias{LinePickingVar}
\title{Get variance for given mode and parameters}
\usage{
  LinePickingVar(mode = 0, para)
}
\arguments{
  \item{mode}{}

  \item{para}{the parameter necessary to describe the space
  given by mode.}
}
\value{
  variance
}
\description{
  For a given mode and parameters gives the variance
}
\note{
  August 25 2012
}
\examples{
LinePickingVar(mode=0,para=10)
}
\author{
  Eric Parsonage, Matt Roughan, Jono Tuke
}

