\docType{methods}
\name{print}
\alias{print}
\alias{print,linepicking-method}
\title{Print the linepicking class to standard output.}
\arguments{
  \item{object}{A \code{linepicking} object.}
}
\value{
  NULL. Prints to standard out.
}
\description{
  Print the linepicking class to standard output.
}
\examples{
tmp <- CreateLinePicking(mode=0,para=10)
tmp
}
\seealso{
  \code{\link{print}}
}

